changed models.py goal_distance_miles to db.Float

+++ models.py
@@ -252,7 +252,7 @@ class Goal(db.Model):
     )
 
     goal_distance_miles = db.Column(
-        db.Integer,
+        db.Float,
         nullable=True
     )
